refactor(main): Change the main packages to Runnable to call from main package
