+++ cmd/server/main.go
@@ -1,11 +1,11 @@
-package main
+package server
 
 import (
 	"container-dsh/pkg/http"
 	"log"
 )
 
-func main() {
+func Run() {
 	if err := http.Run(); err != nil {
 		log.Fatalln("Failed to start server:", err)
 	}
